requirements.txt

streamlit>=1.46.1
numpy>=1.24.0
pandas>=1.5.0
plotly>=5.15.0
Pillow>=9.5.0


altair>=4.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.2.0

opencv-python-headless>=4.8.0

requests>=2.28.0

pickle-mixin>=1.0.2

# Development and Debugging
python-dotenv>=1.0.0

# Performance Optimization
numba>=0.57.0

timm>=1.0.16
